+++ benches/dowser.rs
@@ -19,20 +19,11 @@ benches!(
 	Bench::new("dowser::Dowser::from(/usr/share).collect()")
 		.run(|| Dowser::from("/usr/share").collect::<Vec<_>>()),
 
-	Bench::new("dowser::Dowser::from(/usr/share).into_vec()")
-		.run(|| Dowser::from("/usr/share").into_vec()),
-
 	Bench::spacer(),
 
 	Bench::new("dowser::Dowser::from(/usr/share).filter(gz).collect()")
 		.run_seeded(
 			Dowser::from("/usr/share"),
 			|d| d.filter(|p| Some(GZ) == Extension::try_from2(p)).collect::<Vec<_>>()
-		),
-
-	Bench::new("dowser::Dowser::from(/usr/share).into_vec_filtered(gz)")
-		.run_seeded(
-			Dowser::from("/usr/share"),
-			|d| d.into_vec_filtered(|p| Some(GZ) == Extension::try_from2(p))
 		),
 );